Absorb runtime adapters into app runtime
