+++ crates/perfgate-app/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "perfgate-app"
-description = "Application use-cases and renderers for perfgate workflows"
+description = "Application use-cases, runtime adapters, and renderers for perfgate workflows"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -16,7 +16,6 @@ categories = ["development-tools"]
 [dependencies]
 perfgate-types.workspace = true
 perfgate-domain.workspace = true
-perfgate-adapters.workspace = true
 anyhow.workspace = true
 glob.workspace = true
 handlebars = "6.3"
@@ -26,6 +25,16 @@ serde_json.workspace = true
 time.workspace = true
 toml.workspace = true
 uuid.workspace = true
+thiserror.workspace = true
+hostname = "0.4"
+num_cpus = "1"
+which = "7"
+
+[target.'cfg(windows)'.dependencies]
+windows = { version = "0.60", features = ["Win32_Foundation", "Win32_System_ProcessStatus", "Win32_System_SystemInformation", "Win32_System_Threading"] }
+
+[target.'cfg(unix)'.dependencies]
+libc.workspace = true
 
 [dev-dependencies]
 insta.workspace = true
