+++ crates/perfgate-fake/src/process_runner.rs
@@ -19,7 +19,7 @@ use std::sync::{Arc, Mutex};
 ///
 /// ```
 /// use perfgate_fake::FakeProcessRunner;
-/// use perfgate_adapters::{ProcessRunner, CommandSpec, RunResult};
+/// use perfgate_app::runtime::{ProcessRunner, CommandSpec, RunResult};
 ///
 /// let runner = FakeProcessRunner::new();
 ///
