+++ crates/perfgate/Cargo.toml
@@ -16,7 +16,6 @@ categories.workspace = true
 [dependencies]
 perfgate-types.workspace = true
 perfgate-domain.workspace = true
-perfgate-adapters.workspace = true
 perfgate-app.workspace = true
 anyhow.workspace = true
 regex.workspace = true
