+++ crates/perfgate/src/runtime.rs
@@ -2,3 +2,5 @@
 //! public crates.
 
 pub mod profile;
+
+pub use perfgate_app::runtime::*;
