+++ docs/WORKSPACE.md
@@ -10,6 +10,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 | `perfgate-export` | Workspace-only compatibility wrapper for perfgate::presentation::export | 90% |
 | `perfgate-render` | Workspace-only compatibility wrapper for perfgate::presentation::render | 90% |
 | `perfgate-sensor` | Workspace-only compatibility wrapper for perfgate::presentation::sensor | 90% |
+| `perfgate-adapters` | Workspace-only compatibility wrapper for perfgate::runtime | 90% |
 | `perfgate-github` | Workspace-only compatibility wrapper for perfgate::integrations::github | 90% |
 | `perfgate-paired` | Paired benchmarking statistics (A/B testing) | 100% |
 | `perfgate-fake` | Test utilities and fake implementations | 70% |
@@ -20,8 +21,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 |-------|-------------|------------------|
 | `perfgate-types` | Receipt/config structs, JSON schema types | 95% |
 | `perfgate-domain` | Pure math/policy (I/O-free) | 100% |
-| `perfgate-adapters` | Platform I/O (process execution, host probing) | 80% |
-| `perfgate-app` | Use-cases, orchestration layer | 90% |
+| `perfgate-app` | Use-cases, runtime adapters, orchestration layer | 90% |
 | `perfgate-cli` | CLI argument parsing and command dispatch | 70% |
 | `perfgate-server` | REST API server for baseline management | 90% |
 | `perfgate-client` | API client for baseline server interaction | 90% |
@@ -39,8 +39,8 @@ graph TD
   domain --> budget[perfgate-domain::budget]
   domain --> sig[perfgate-domain::significance]
   domain --> scaling[perfgate-domain::scaling]
-  domain --> adapters[perfgate-adapters]
-  adapters --> app[perfgate-app]
+  domain --> runtime[perfgate-app::runtime]
+  runtime --> app[perfgate-app]
   app --> cli[perfgate-cli]
   app --> facade[perfgate]
   types --> client[perfgate-client]
