+++ policy/absorbed_crates.txt
@@ -34,7 +34,7 @@ perfgate-export -> perfgate::presentation::export [compatibility wrapper] [publi
 perfgate-sensor -> perfgate::presentation::sensor [compatibility wrapper] [publish=false]
 
 # Runtime, app, and integrations.
-perfgate-adapters -> perfgate::runtime
+perfgate-adapters -> perfgate_app::runtime; perfgate::runtime facade [compatibility wrapper] [publish=false]
 perfgate-profile -> perfgate::runtime::profile [deleted]
 perfgate-app -> perfgate::app
 perfgate-github -> perfgate::integrations::github [compatibility wrapper] [publish=false]
